# nra-tax-engine/src/api/main.py
"""FastAPI surface.

Phase 6: switches the primary submission endpoint to accept a typed
:class:`IntakePayload` instead of a free-form JSON string. The OpenAPI
schema auto-generated by FastAPI now describes every intake field, so the
client can codegen its TypeScript types from ``/openapi.json``.

The legacy multipart endpoint was removed (dead code, unnecessary file-upload
attack surface). New clients should use ``POST /api/v1/submit`` with the typed
:class:`SubmitRequest` body or ``POST /api/v1/ocr`` for document extraction.
"""

from __future__ import annotations

import logging
import os
import uuid
from pathlib import Path
from typing import List, Optional

import uvicorn
from fastapi import Depends, FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import FileResponse
from pydantic import BaseModel, Field

from src.assembly.mailing_packager import MailingPackager
from src.api.auth import require_api_key
from src.api.ocr_endpoint import router as ocr_router
from src.intake.intake_schema import IntakePayload
from src.intake.mcq_router import MCQRouter
from src.orchestrator.engine import TaxEngine

logger = logging.getLogger(__name__)


def _handle_pipeline_error(exc: Exception, correlation: str) -> None:
    """Log a pipeline failure server-side and raise a generic 500.

    The raw exception text (which can contain file paths, stack frames, or
    PII from the intake payload) must never reach the client. We log the full
    detail locally and surface only an opaque correlation id.
    """
    logger.exception("Pipeline failure [%s]: %s", correlation, exc)
    raise HTTPException(
        status_code=500,
        detail=f"Internal processing error. Reference: {correlation}",
    ) from exc


class TaxProcessResponse(BaseModel):
    """Top-level API response."""

    status: str = Field(description="'success' or 'error'.")
    tax_year: int
    federal_refund_or_owed: float = Field(
        description="Positive = owes IRS; negative = federal refund due."
    )
    ny_refund_or_owed: float = Field(
        default=0.0,
        description="NY positive = owes NY DTF; negative = NY refund.",
    )
    fica_refund_amount: float = Field(
        default=0.0, description="Total Form 843 FICA refund amount (Ssec + Medicare)."
    )
    forms_required: List[str] = Field(default_factory=list)
    completed_layers: List[str] = Field(default_factory=list)
    generated_form_outputs: List[str] = Field(
        default_factory=list,
        description="Paths to per-form filled PDFs or field-map JSONs.",
    )
    federal_packet_path: Optional[str] = None
    ny_packet_path: Optional[str] = None
    fica_packet_path: Optional[str] = None
    requires_human_review: List[str] = Field(
        default_factory=list,
        description=(
            "Reasons a CPA must review the return before filing. Non-empty "
            "when a validator or LLM confidence check flagged something. "
            "Pass force_assembly=True on a subsequent call to override after review."
        ),
    )
    narrative_sections: dict = Field(
        default_factory=dict,
        description=(
            "Plain-English narrative sections keyed by section title. "
            "Each value is a human-readable explanation of that portion of the return. "
            "Empty dict when narrative generation is not available."
        ),
    )


app = FastAPI(
    title="QuadTax NRA Tax Engine API",
    version="0.4.0",
    description=(
        "Production NRA tax return engine. Accepts intake fields and "
        "OCR'd document text; returns refund/owed amounts and paths to "
        "generated forms and Pub-519-Ch-8-ordered mailing packets."
    ),
)

# CORS: origins are environment-driven so production can pin specific frontends
# instead of the localhost dev defaults. Methods/headers are restricted to what
# the API actually uses — `*` with allow_credentials=True is too broad for a
# system handling SSN/ITIN PII.
_CORS_ORIGINS = [
    o.strip()
    for o in os.getenv(
        "QUADTAX_CORS_ORIGINS",
        "http://localhost:3000,http://127.0.0.1:3000",
    ).split(",")
    if o.strip()
]

app.add_middleware(
    CORSMiddleware,
    allow_origins=_CORS_ORIGINS,
    allow_credentials=True,
    allow_methods=["GET", "POST"],
    allow_headers=["Content-Type", "Authorization"],
)


@app.middleware("http")
async def security_headers(request, call_next):
    """Add baseline security headers to every response."""
    response = await call_next(request)
    response.headers.setdefault("X-Content-Type-Options", "nosniff")
    response.headers.setdefault("X-Frame-Options", "DENY")
    response.headers.setdefault("Referrer-Policy", "no-referrer")
    response.headers.setdefault(
        "Permissions-Policy", "geolocation=(), microphone=(), camera=()"
    )
    return response

app.include_router(ocr_router)


# ---------------------------------------------------------------------------
# Modern endpoint: typed JSON payload
# ---------------------------------------------------------------------------


class SubmitRequest(BaseModel):
    """Inputs to ``POST /api/v1/submit``."""

    intake: IntakePayload
    i94_ocr_text: str = ""
    w2_ocr_texts: List[str] = Field(default_factory=list)
    form_1042s_ocr_texts: List[str] = Field(default_factory=list)
    output_dir: str = Field(default="outputs", description="Directory to write packets into.")
    force_assembly: bool = Field(
        default=False,
        description=(
            "When True the engine bypasses the human-review gate. The caller "
            "is acknowledging every reason previously surfaced in "
            "``requires_human_review``. Default False — first request returns "
            "the review list without producing forms."
        ),
    )


@app.get("/api/v1/healthz", tags=["meta"])
def healthz() -> dict:
    return {"status": "ok"}


@app.post(
    "/api/v1/submit",
    response_model=TaxProcessResponse,
    tags=["tax"],
    dependencies=[Depends(require_api_key)],
)
def submit(request: SubmitRequest) -> TaxProcessResponse:
    """Run the full L1-L9 pipeline against a typed intake payload."""
    router = MCQRouter()
    mcq_dict = router.to_mcq_answers(request.intake)

    # Pre-seed identity so Phase-3 add-ons (AMT, ITIN, Form 2210) read the
    # right values when they run inside run_full_pipeline().
    seeded_state = router.populate_state(request.intake)

    engine = TaxEngine(force_assembly=request.force_assembly)
    correlation = uuid.uuid4().hex[:12]
    try:
        generated_forms, state = engine.run_full_pipeline(
            i94_ocr_text=request.i94_ocr_text,
            w2_ocr_texts=request.w2_ocr_texts,
            form_1042s_ocr_texts=request.form_1042s_ocr_texts,
            mcq_answers=mcq_dict,
            initial_state=seeded_state,
        )
    except Exception as exc:  # noqa: BLE001
        _handle_pipeline_error(exc, correlation)

    # Build mailing packets from whatever the populator wrote into outputs/.
    packager = MailingPackager(tax_year=state.tax_year)
    package = packager.assemble(
        state,
        forms_dir=Path("outputs"),
        output_dir=Path(request.output_dir),
    )

    from src.narrative.generator import NarrativeGenerator
    narrative = NarrativeGenerator().generate_sections(state)

    fica_total = (
        state.fica.incorrect_ss_withheld + state.fica.incorrect_medicare_withheld
    )
    return TaxProcessResponse(
        status="success",
        tax_year=state.tax_year,
        federal_refund_or_owed=float(state.tax.refund_or_owed),
        ny_refund_or_owed=float(state.ny.ny_refund_or_owed),
        fica_refund_amount=float(fica_total),
        forms_required=list(state.forms_required),
        completed_layers=list(state.completed_layers),
        generated_form_outputs=list(generated_forms),
        federal_packet_path=package.federal.pdf_output or package.federal.json_output,
        ny_packet_path=(
            package.ny.pdf_output or package.ny.json_output if package.ny else None
        ),
        fica_packet_path=(
            package.fica_843.pdf_output or package.fica_843.json_output
            if package.fica_843
            else None
        ),
        requires_human_review=list(state.requires_human_review),
        narrative_sections=narrative,
    )


# ---------------------------------------------------------------------------
# Note: the legacy multipart "/api/v1/upload-and-process" endpoint was removed.
# It was dead code (never called by the Phase-6 client) and an unnecessary
# file-upload attack surface for a system handling SSN/ITIN PII. New clients
# should use ``POST /api/v1/submit`` with the typed :class:`SubmitRequest` body
# or ``POST /api/v1/ocr`` for document extraction.
# ---------------------------------------------------------------------------


@app.get(
    "/api/v1/packet",
    tags=["tax"],
    dependencies=[Depends(require_api_key)],
)
def download_packet(path: str) -> FileResponse:
    """Serve a generated packet file. Only allows files under outputs/ directory."""
    abs_path = os.path.realpath(path)
    outputs_abs = os.path.realpath("outputs")
    # Strict containment: must equal outputs_abs or live in a sub-directory of
    # it. A bare startswith() check is bypassable by paths like
    # "outputs_evil/file.pdf", which is a prefix of "outputs/".
    if abs_path != outputs_abs and not abs_path.startswith(outputs_abs + os.sep):
        raise HTTPException(status_code=403, detail="Access denied.")
    if not os.path.exists(abs_path):
        raise HTTPException(status_code=404, detail="Packet not found.")
    return FileResponse(abs_path, media_type="application/pdf",
                        filename=os.path.basename(abs_path))


if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8000)